discontinue vapoursynth support and simplify the formula
This is the first step in closing down this repo. I removed Vapoursynth since
it now only has fringe use cases. We now have support for high quality
debanding using shaders, and you can use more advanced features for frame
timing than when we added mvtools support to the formula.

With that in mind the code complexity and pain of dealing with it in the mpv
formula was not worth it anymore (as you can see the new code is quite
smaller!).

If no bugs are found, I'll open a PR with this new simpler formula to
homebrew's official repository in a week or so, and close down the repo!

More info at Homebrew/homebrew#46413